vc-chlog: new option --dirty-workdir for git add --interactive.
With 'git add -i', it can be helpful to pipe the output of
'git diff --cached' into 'vc-chlog --stdin'.  However, if the
working directory also contains unrelated changes, then the tag
list might be wrong, because we try to recreate the old version
of the file from the copy in the working directory.  With the
new --dirty-workdir options, use 'vc-dwim --diff' to get back to
the state of the last commit.  From there, recreate the version
of the new file with the patch from stdin.

* vc-chlog.in: New option --dirty-workdir.
(func_help): Document it.
(func_extract_identifiers): If the work directory is dirty,
recreate the old file from 'vc-dwim --diff' rather than from the
patch in question; also, create a version of the new file based
on that plus the patch in question.
* doc/vc-dwim.texi (vc-chlog Invocation): Document the option.
* tests/dirty-workdir: Use it.
* tests/Makefile.am (XFAIL_TESTS): Remove dirty-workdir.
* NEWS: Update.

Signed-off-by: Ralf Wildenhues <Ralf.Wildenhues@gmx.de>