maint: enable running from srcdir prior to make install
Running ./vc-dwim would fail when run pre-"make-install".  Of course,
one could inspect the source or tests and find the trick of setting the
perllibdir envvar, but one should not have to do that.
* vc-dwim.pl (BEGIN): Prepend dirname $0 to @INC.
This also makes it so a manual post-install test uses .pm files
from the current directory rather than the installed ones.
* Makefile.am (install-exec-hook): Upon installation, remove the line
that prepends basename $0 to @INC
Reported by Karl Berry.